+++ src/DSF/DSFPointPool.cpp
@@ -314,7 +314,7 @@ pair<int, int>	DSFContiguousPointPool::AccumulatePoints(const DSFTupleVector& in
 			return pair<int, int>(p, pos);
 		}
 	}
-	printf("point pool failure...%d out of %d tried.  Points = %d.\n", tris, p, inPoints.size());
+	printf("point pool failure...%d out of %d tried.  Points = %llu.\n", tris, p, (unsigned long long)inPoints.size());
 		inPoints[n].dump();
 	printf("\n");
 	return pair<int, int>(-1, -1);
